Move reputation types from reth-network-api to reth-network-types
