+++ crates/net/network/src/fetch/client.rs
@@ -3,7 +3,6 @@
 use crate::{fetch::DownloadRequest, flattened_response::FlattenedResponse, peers::PeersHandle};
 use futures::{future, future::Either};
 
-use reth_network_api::ReputationChangeKind;
 use reth_network_p2p::{
     bodies::client::{BodiesClient, BodiesFut},
     download::DownloadClient,
@@ -12,6 +11,7 @@ use reth_network_p2p::{
     priority::Priority,
 };
 use reth_network_peers::PeerId;
+use reth_network_types::ReputationChangeKind;
 use reth_primitives::{Header, B256};
 use std::sync::{
     atomic::{AtomicUsize, Ordering},
