+++ crates/net/network/src/fetch/mod.rs
@@ -3,13 +3,13 @@
 use crate::{message::BlockRequest, peers::PeersHandle};
 use futures::StreamExt;
 use reth_eth_wire::{GetBlockBodies, GetBlockHeaders};
-use reth_network_api::ReputationChangeKind;
 use reth_network_p2p::{
     error::{EthResponseValidator, PeerRequestResult, RequestError, RequestResult},
     headers::client::HeadersRequest,
     priority::Priority,
 };
 use reth_network_peers::PeerId;
+use reth_network_types::ReputationChangeKind;
 use reth_primitives::{BlockBody, Header, B256};
 use std::{
     collections::{HashMap, VecDeque},
