+++ crates/net/p2p/Cargo.toml
@@ -13,11 +13,11 @@ workspace = true
 
 [dependencies]
 reth-primitives.workspace = true
-reth-network-api.workspace = true
 reth-eth-wire-types.workspace = true
 reth-consensus.workspace = true
 reth-network-peers.workspace = true
 reth-storage-errors.workspace = true
+reth-network-types.workspace = true
 
 # async
 futures.workspace = true
