+++ crates/net/p2p/src/error.rs
@@ -1,7 +1,7 @@
 use super::headers::client::HeadersRequest;
 use reth_consensus::ConsensusError;
-use reth_network_api::ReputationChangeKind;
 use reth_network_peers::WithPeerId;
+use reth_network_types::ReputationChangeKind;
 use reth_primitives::{
     BlockHashOrNumber, BlockNumber, GotExpected, GotExpectedBoxed, Header, B256,
 };
